requirements.txt
certifi==2020.12.5
chainer==7.7.0
chumpy==0.70
cupy-cuda102==7.7.0
cycler==0.10.0
decorator==4.4.2
easydict==1.9
fastrlock==0.6
filelock==3.0.12
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy==1.20.2
olefile==0.46
opencv-python==4.5.1.48
Pillow @ file:///tmp/build/80754af9/pillow_1615057347459/work
protobuf==3.15.6
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.2
six @ file:///tmp/build/80754af9/six_1605205313296/work
tifffile==2021.3.17
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tqdm==4.59.0
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
